+++ src/entities/errors.ts
@@ -62,7 +62,7 @@ export class ErrorCodeUtils {
       case ErrorCode.MissingReceiptFileError:
         return "The receipt is missing.";
       case ErrorCode.NetworkError:
-        return "Error performing request.";
+        return "Error performing request. Please check your network connection and try again.";
       case ErrorCode.InvalidCredentialsError:
         return "There was a credentials issue. Check the underlying error for more details.";
       case ErrorCode.UnexpectedBackendResponseError:
@@ -112,6 +112,7 @@ export class ErrorCodeUtils {
   ): ErrorCode {
     switch (backendErrorCode) {
       case BackendErrorCode.BackendStoreProblem:
+      case BackendErrorCode.BackendPaymentGatewayGenericError:
         return ErrorCode.StoreProblemError;
       case BackendErrorCode.BackendCannotTransferPurchase:
         return ErrorCode.ReceiptAlreadyInUseError;
@@ -208,8 +209,24 @@ export enum BackendErrorCode {
   BackendInvalidSubscriberAttributesBody = 7264,
   BackendProductIDsMalformed = 7662,
   BackendAlreadySubscribedError = 7772,
+  BackendPaymentGatewayGenericError = 7773,
   BackendOfferNotFound = 7814,
   BackendNoMXRecordsFound = 7834,
+}
+
+/**
+ * Extra information that is available in certain types of errors.
+ * @public
+ */
+export interface PurchasesErrorExtra {
+  /**
+   * If this is a request error, the HTTP status code of the response.
+   */
+  readonly statusCode?: number;
+  /**
+   * If this is a RevenueCat backend error, the error code from the servers.
+   */
+  readonly backendErrorCode?: number;
 }
 
 /**
@@ -229,6 +246,7 @@ export class PurchasesError extends Error {
       errorCode,
       ErrorCodeUtils.getPublicMessage(errorCode),
       backendErrorMessage,
+      { backendErrorCode: backendErrorCode },
     );
   }
 
@@ -260,6 +278,10 @@ export class PurchasesError extends Error {
      * can be useful for debugging and logging.
      */
     public readonly underlyingErrorMessage?: string | null,
+    /**
+     * Contains extra information that is available in certain types of errors.
+     */
+    public readonly extra?: PurchasesErrorExtra,
   ) {
     super(message);
   }