+++ src/networking/backend.ts
@@ -6,10 +6,10 @@ import {
   GetCustomerInfoEndpoint,
   GetOfferingsEndpoint,
   GetProductsEndpoint,
-  SubscribeEndpoint,
+  PurchaseEndpoint,
 } from "./endpoints";
 import { type SubscriberResponse } from "./responses/subscriber-response";
-import { type SubscribeResponse } from "./responses/subscribe-response";
+import { type PurchaseResponse } from "./responses/purchase-response";
 import { type ProductsResponse } from "./responses/products-response";
 import { type BrandingInfoResponse } from "./responses/branding-response";
 import { type CheckoutStatusResponse } from "./responses/checkout-status-response";
@@ -72,18 +72,19 @@ export class Backend {
     );
   }
 
-  async postSubscribe(
+  async postPurchase(
     appUserId: string,
     productId: string,
     email: string,
     presentedOfferingContext: PresentedOfferingContext,
     purchaseOption: PurchaseOption,
-  ): Promise<SubscribeResponse> {
-    type SubscribeRequestBody = {
+  ): Promise<PurchaseResponse> {
+    type PurchaseRequestBody = {
       app_user_id: string;
       product_id: string;
       email: string;
       presented_offering_identifier: string;
+      presented_placement_identifier?: string;
       offer_id?: string;
       price_id: string;
       applied_targeting_rule?: {
@@ -92,7 +93,7 @@ export class Backend {
       };
     };
 
-    const requestBody: SubscribeRequestBody = {
+    const requestBody: PurchaseRequestBody = {
       app_user_id: appUserId,
       product_id: productId,
       email: email,
@@ -112,8 +113,13 @@ export class Backend {
       };
     }
 
-    return await performRequest<SubscribeRequestBody, SubscribeResponse>(
-      new SubscribeEndpoint(),
+    if (presentedOfferingContext.placementIdentifier) {
+      requestBody.presented_placement_identifier =
+        presentedOfferingContext.placementIdentifier;
+    }
+
+    return await performRequest<PurchaseRequestBody, PurchaseResponse>(
+      new PurchaseEndpoint(),
       {
         apiKey: this.API_KEY,
         body: requestBody,