+++ src/tests/networking/backend.test.ts
@@ -14,7 +14,7 @@ import {
   PurchasesError,
 } from "../../entities/errors";
 import { expectPromiseToError } from "../test-helpers";
-import { type SubscribeResponse } from "../../networking/responses/subscribe-response";
+import { type PurchaseResponse } from "../../networking/responses/purchase-response";
 
 let server: SetupServer;
 let backend: Backend;
@@ -108,7 +108,8 @@ describe("getCustomerInfo request", () => {
       backend.getCustomerInfo("someAppUserId"),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        "Unknown backend error. Request: getCustomerInfo. Status code: 500. Body: null.",
+        "Unknown backend error.",
+        "Request: getCustomerInfo. Status code: 500. Body: null.",
       ),
     );
   });
@@ -147,7 +148,8 @@ describe("getCustomerInfo request", () => {
       backend.getCustomerInfo("someAppUserId"),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        'Unknown backend error. Request: getCustomerInfo. Status code: 400. Body: {"code":1234567890,"message":"Invalid error message"}.',
+        "Unknown backend error.",
+        'Request: getCustomerInfo. Status code: 400. Body: {"code":1234567890,"message":"Invalid error message"}.',
       ),
     );
   });
@@ -160,7 +162,8 @@ describe("getCustomerInfo request", () => {
       backend.getCustomerInfo("someAppUserId"),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        "Unknown backend error. Request: getCustomerInfo. Status code: 400. Body: null.",
+        "Unknown backend error.",
+        "Request: getCustomerInfo. Status code: 400. Body: null.",
       ),
     );
   });
@@ -171,7 +174,7 @@ describe("getCustomerInfo request", () => {
       backend.getCustomerInfo("someAppUserId"),
       new PurchasesError(
         ErrorCode.NetworkError,
-        "Error performing request.",
+        "Error performing request. Please check your network connection and try again.",
         "Failed to fetch",
       ),
     );
@@ -209,7 +212,8 @@ describe("getOfferings request", () => {
       backend.getOfferings("someAppUserId"),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        "Unknown backend error. Request: getOfferings. Status code: 500. Body: null.",
+        "Unknown backend error.",
+        "Request: getOfferings. Status code: 500. Body: null.",
       ),
     );
   });
@@ -240,7 +244,7 @@ describe("getOfferings request", () => {
       backend.getOfferings("someAppUserId"),
       new PurchasesError(
         ErrorCode.NetworkError,
-        "Error performing request.",
+        "Error performing request. Please check your network connection and try again.",
         "Failed to fetch",
       ),
     );
@@ -298,7 +302,8 @@ describe("getProducts request", () => {
       backend.getProducts("someAppUserId", ["monthly", "monthly_2"]),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        "Unknown backend error. Request: getProducts. Status code: 500. Body: null.",
+        "Unknown backend error.",
+        "Request: getProducts. Status code: 500. Body: null.",
       ),
     );
   });
@@ -329,63 +334,72 @@ describe("getProducts request", () => {
       backend.getProducts("someAppUserId", ["monthly", "monthly_2"]),
       new PurchasesError(
         ErrorCode.NetworkError,
-        "Error performing request.",
+        "Error performing request. Please check your network connection and try again.",
         "Failed to fetch",
       ),
     );
   });
 });
 
-describe("subscribe request", () => {
-  function setSubscribeResponse(httpResponse: HttpResponse) {
+describe("purchase request", () => {
+  function setPurchaseResponse(httpResponse: HttpResponse) {
     server.use(
-      http.post("http://localhost:8000/rcbilling/v1/subscribe", () => {
+      http.post("http://localhost:8000/rcbilling/v1/purchase", () => {
         return httpResponse;
       }),
     );
   }
 
-  test("can post subscribe successfully", async () => {
-    const subscribeResponse: SubscribeResponse = {
+  test("can post purchase successfully", async () => {
+    const purchaseResponse: PurchaseResponse = {
       operation_session_id: "test-operation-session-id",
       next_action: "collect_payment_info",
       data: {
         client_secret: "seti_123",
       },
     };
-    setSubscribeResponse(HttpResponse.json(subscribeResponse, { status: 200 }));
+    setPurchaseResponse(HttpResponse.json(purchaseResponse, { status: 200 }));
     expect(
-      await backend.postSubscribe(
+      await backend.postPurchase(
         "someAppUserId",
         "monthly",
         "testemail@revenuecat.com",
-        { offeringIdentifier: "offering_1", targetingContext: null },
+        {
+          offeringIdentifier: "offering_1",
+          targetingContext: null,
+          placementIdentifier: null,
+        },
         { id: "base_option", priceId: "test_price_id" },
       ),
-    ).toEqual(subscribeResponse);
+    ).toEqual(purchaseResponse);
   });
 
   test("throws an error if the backend returns a server error", async () => {
-    setSubscribeResponse(
+    setPurchaseResponse(
       HttpResponse.json(null, { status: StatusCodes.INTERNAL_SERVER_ERROR }),
     );
     await expectPromiseToError(
-      backend.postSubscribe(
+      backend.postPurchase(
         "someAppUserId",
         "monthly",
         "testemail@revenuecat.com",
-        { offeringIdentifier: "offering_1", targetingContext: null },
+        {
+          offeringIdentifier: "offering_1",
+          targetingContext: null,
+          placementIdentifier: null,
+        },
         { id: "base_option", priceId: "test_price_id" },
       ),
       new PurchasesError(
         ErrorCode.UnknownBackendError,
-        "Unknown backend error. Request: subscribe. Status code: 500. Body: null.",
+        "Unknown backend error.",
+        "Request: purchase. Status code: 500. Body: null.",
       ),
     );
   });
 
   test("throws a known error if the backend returns a request error with correct body", async () => {
-    setSubscribeResponse(
+    setPurchaseResponse(
       HttpResponse.json(
         {
           code: BackendErrorCode.BackendInvalidAPIKey,
@@ -395,11 +409,15 @@ describe("subscribe request", () => {
       ),
     );
     await expectPromiseToError(
-      backend.postSubscribe(
+      backend.postPurchase(
         "someAppUserId",
         "monthly",
         "testemail@revenuecat.com",
-        { offeringIdentifier: "offering_1", targetingContext: null },
+        {
+          offeringIdentifier: "offering_1",
+          targetingContext: null,
+          placementIdentifier: null,
+        },
         { id: "base_option", priceId: "test_price_id" },
       ),
       new PurchasesError(
@@ -411,7 +429,7 @@ describe("subscribe request", () => {
   });
 
   test("throws a PurchaseInvalidError if the backend returns with a offer not found error", async () => {
-    setSubscribeResponse(
+    setPurchaseResponse(
       HttpResponse.json(
         {
           code: BackendErrorCode.BackendOfferNotFound,
@@ -421,11 +439,15 @@ describe("subscribe request", () => {
       ),
     );
     await expectPromiseToError(
-      backend.postSubscribe(
+      backend.postPurchase(
         "someAppUserId",
         "monthly",
         "testemail@revenuecat.com",
-        { offeringIdentifier: "offering_1", targetingContext: null },
+        {
+          offeringIdentifier: "offering_1",
+          targetingContext: null,
+          placementIdentifier: null,
+        },
         { id: "base_option", priceId: "test_price_id" },
       ),
       new PurchasesError(
@@ -437,18 +459,22 @@ describe("subscribe request", () => {
   });
 
   test("throws network error if cannot reach server", async () => {
-    setSubscribeResponse(HttpResponse.error());
+    setPurchaseResponse(HttpResponse.error());
     await expectPromiseToError(
-      backend.postSubscribe(
+      backend.postPurchase(
         "someAppUserId",
         "monthly",
         "testemail@revenuecat.com",
-        { offeringIdentifier: "offering_1", targetingContext: null },
+        {
+          offeringIdentifier: "offering_1",
+          targetingContext: null,
+          placementIdentifier: null,
+        },
         { id: "base_option", priceId: "test_price_id" },
       ),
       new PurchasesError(
         ErrorCode.NetworkError,
-        "Error performing request.",
+        "Error performing request. Please check your network connection and try again.",
         "Failed to fetch",
       ),
     );