+++ src/ui/rcb-ui.svelte
@@ -9,7 +9,7 @@
   import StateNeedsAuthInfo from "./states/state-needs-auth-info.svelte";
   import ConditionalFullScreen from "./conditional-full-screen.svelte";
   import Shell from "./shell.svelte";
-  import { type SubscribeResponse } from "../networking/responses/subscribe-response";
+  import { type PurchaseResponse } from "../networking/responses/purchase-response";
   import { type BrandingInfoResponse } from "../networking/responses/branding-response";
   import {
     PurchaseFlowError,
@@ -35,7 +35,7 @@
 
   let colorVariables = "";
   let productDetails: Product | null = null;
-  let paymentInfoCollectionMetadata: SubscribeResponse | null = null;
+  let paymentInfoCollectionMetadata: PurchaseResponse | null = null;
   let lastError: PurchaseFlowError | null = null;
   const productId = rcPackage.rcBillingProduct.identifier ?? null;
   const defaultPurchaseOption =
@@ -164,7 +164,7 @@
   };
 
   const handleError = (e: PurchaseFlowError) => {
-    if (state === "processing-auth-info") {
+    if (state === "processing-auth-info" && e.isRecoverable()) {
       lastError = e;
       state = "needs-auth-info";
       return;
@@ -253,7 +253,7 @@
             />
           {/if}
           {#if state === "success"}
-            <StateSuccess {brandingInfo} onContinue={handleContinue} />
+            <StateSuccess {productDetails} {brandingInfo} onContinue={handleContinue} />
           {/if}
         </Shell>
       </div>