+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## 0.8.0
+## RevenueCat SDK
+### ✨ New Features
+* Add support for `placements` (#189) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* Bump fastlane from 2.222.0 to 2.223.1 (#192) via dependabot[bot] (@dependabot[bot])
+* Bumped `vite-plugin-dts` version to the latest one (#184) via Nicola Sacco (@nicfix)
+* Bump danger from 9.4.3 to 9.5.0 (#188) via dependabot[bot] (@dependabot[bot])
+* Bump vite from 5.4.2 to 5.4.6 (#185) via dependabot[bot] (@dependabot[bot])
+* Bump body-parser and express (#179) via dependabot[bot] (@dependabot[bot])
+
+### 🔄 Other Changes
+* Update fastlane plugin (#195) via Toni Rico (@tonidero)
+* Update vite in demo app (#190) via Toni Rico (@tonidero)
+* Add additional error details and display (#186) via Toni Rico (@tonidero)
+* Create dependabot.yml (#187) via Toni Rico (@tonidero)
+* [AUTOMATIC] Release/0.7.0 (#175) via RevenueCat Git Bot (@RCGitBot)
+
+## 0.7.0
+### New Features
+* Create a `preload` method in `Purchases` to prefetch branding information. (#169) via Toni Rico (@tonidero)
+* Add `amountMicros` to `Price` interface (#170) via Toni Rico (@tonidero)
+### Bugfixes
+* Improve purchase flow error handling for unknown errors (#173) via Toni Rico (@tonidero)
+### Dependency Updates
+* Bump svelte from 4.2.12 to 4.2.19 (#168) via dependabot[bot] (@dependabot[bot])
+### Other Changes
+* ECO-1398: Added the possibility to pass a currency in the demo app query string params (#167) via Nicola Sacco (@nicfix)
+
 ## 0.6.0
 ### New Features
 * [Beta] Support different currencies (#161) via Toni Rico (@tonidero)